+++ _website/.yarnrc.yml
@@ -1,1 +1,1 @@
-nodeLinker: node-modules
+nodeLinker: pnpm
