+++ _website/src/components/detail/Items.tsx
@@ -1,4 +1,4 @@
-import { VersionData } from '../../utils/parser/types';
+import { VersionData } from '@/utils/parser/types';
 
 interface VersionItemProps {
     ver: VersionData[string]['versions'][number];
