+++ _website/vite.config.ts
@@ -6,4 +6,9 @@ import tailwindcss from '@tailwindcss/vite';
 export default defineConfig({
     base: './',
     plugins: [react(), tailwindcss()],
+    resolve: {
+        alias: {
+            '@': '/src',
+        },
+    },
 });
